Updated deploy-book.yml
$GITHUB_PATH: Used $(pwd)/mdbook/bin instead of just pwd to correctly add the mdbook binary to the environment path.
Email configuration: Replaced the empty email in the git config with a placeholder "deploy@example.com". You can replace it with a proper email.
Directory creation: Used mkdir -p mdbook to ensure that the directory is created only if it doesn't already exist.
Additional Comments: Added some comments to clarify the steps for future maintainers.